requirements.txt
python-socketio[client]
requests
flask
